+++ src/jamma/pipeline_memory.py
@@ -8,7 +8,8 @@ from jamma.core import memory
 from jamma.core.eigen_plan import EigenDriverPlan
 from jamma.lmm.eigen import plan_eigen_driver_for_machine
 from jamma.lmm.loco_eigen import plan_loco_eigen_driver
-from jamma.pipeline_plan import AnalysisPlan, LocoAnalysisPlan, ProvidedEigen
+from jamma.pipeline_config import ProvidedEigen
+from jamma.pipeline_plan import AnalysisPlan, LocoAnalysisPlan
 
 __all__ = ["memory_preflight"]